# apps/web/src/lib/api/generated/openapi.json
{
  "components": {
    "schemas": {
...
      "GSTRequest": {
        "properties": {
          "is_interstate": {
            "title": "Is Interstate",
            "type": "boolean"
          },
          "rate_percent": {
            "title": "Rate Percent",
            "type": "number"
          },
          "taxable_paise": {
            "title": "Taxable Paise",
            "type": "integer"
          }
        },
        "required": [
          "taxable_paise",
          "rate_percent",
          "is_interstate"
        ],
        "title": "GSTRequest",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ITCEligibilityRequest": {
        "properties": {
          "gstr2b_amount_paise": {
            "title": "Gstr2B Amount Paise",
            "type": "integer"
          },
          "invoice_amount_paise": {
            "title": "Invoice Amount Paise",
            "type": "integer"
          },
          "supplier_filed": {
            "title": "Supplier Filed",
            "type": "boolean"
          },
          "tolerance_paise": {
            "default": 100,
            "title": "Tolerance Paise",
            "type": "integer"
          }
        },
        "required": [
          "supplier_filed",
          "invoice_amount_paise",
          "gstr2b_amount_paise"
        ],
        "title": "ITCEligibilityRequest",
        "type": "object"
      },
...
      "InterestRequest": {
        "properties": {
          "actual_filing_date": {
            "format": "date",
            "title": "Actual Filing Date",
            "type": "string"
          },
          "advance_tax_paid_paise": {
            "title": "Advance Tax Paid Paise",
            "type": "integer"
          },
          "assessment_year_end": {
            "format": "date",
            "title": "Assessment Year End",
            "type": "string"
          },
          "installments_paid": {
            "default": [],
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "title": "Installments Paid",
            "type": "array"
          },
          "total_tax_paise": {
            "title": "Total Tax Paise",
            "type": "integer"
          }
        },
        "required": [
          "total_tax_paise",
          "advance_tax_paid_paise",
          "assessment_year_end",
          "actual_filing_date"
        ],
        "title": "InterestRequest",
        "type": "object"
      },
...
      "NetGSTPayableRequest": {
        "properties": {
          "itc_available_paise": {
            "title": "Itc Available Paise",
            "type": "integer"
          },
          "itc_eligible_paise": {
            "title": "Itc Eligible Paise",
            "type": "integer"
          },
          "output_gst_paise": {
            "title": "Output Gst Paise",
            "type": "integer"
          }
        },
        "required": [
          "output_gst_paise",
          "itc_available_paise",
          "itc_eligible_paise"
        ],
        "title": "NetGSTPayableRequest",
        "type": "object"
      },
...
      "TDSRequest": {
        "properties": {
          "gross_amount_paise": {
            "title": "Gross Amount Paise",
            "type": "integer"
          },
          "pan_available": {
            "default": true,
            "title": "Pan Available",
            "type": "boolean"
          },
          "section": {
            "title": "Section",
            "type": "string"
          }
        },
        "required": [
          "section",
          "gross_amount_paise"
        ],
        "title": "TDSRequest",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "title": "firmOS API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/compute/gst": {
      "post": {
        "description": "Compute CGST, SGST, IGST from taxable amount and rate.",
        "operationId": "compute_gst_api_compute_gst_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GSTRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Compute Gst",
        "tags": [
          "compute"
        ]
      }
    },
    "/api/compute/interest": {
      "post": {
        "description": "Compute 234B and 234C interest for advance tax defaults.",
        "operationId": "compute_interest_api_compute_interest_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/InterestRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Compute Interest",
        "tags": [
          "compute"
        ]
      }
    },
    "/api/compute/itc-eligibility": {
      "post": {
        "description": "Check ITC eligibility against GSTR-2B data.",
        "operationId": "check_itc_api_compute_itc_eligibility_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ITCEligibilityRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Check Itc",
        "tags": [
          "compute"
        ]
      }
    },
    "/api/compute/net-gst-payable": {
      "post": {
        "description": "Compute net GST payable after eligible ITC deduction.",
        "operationId": "compute_net_gst_api_compute_net_gst_payable_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NetGSTPayableRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Compute Net Gst",
        "tags": [
          "compute"
        ]
      }
    },
    "/api/compute/tds": {
      "post": {
        "description": "Compute TDS amount based on section and PAN availability.",
        "operationId": "compute_tds_api_compute_tds_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TDSRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Compute Tds",
        "tags": [
          "compute"
        ]
      }
    },
...
    "/api/connector-operations/sync-jobs/{job_id}/retry": {
      "post": {
        "description": "Resume a failed read job from its saved cursor; financial writes are never retried here.",
        "operationId": "retry_sync_job_api_connector_operations_sync_jobs__job_id__retry_post",
        "parameters": [
          {
            "in": "path",
            "name": "job_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Job Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Retry Sync Job",
        "tags": [
          "connector-operations"
        ]
      }
    },
...
    "/api/connectors/{connector_id}/connect": {
      "post": {
        "operationId": "connect_connector_api_connectors__connector_id__connect_post",
        "parameters": [
          {
            "in": "path",
            "name": "connector_id",
            "required": true,
            "schema": {
              "title": "Connector Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Connect Connector",
        "tags": [
          "connectors"
        ]
      }
    },
...
  }
}